refactor. Moved methode for automatic data download in new service class
